+++ backend/src/main/java/io/mixeway/mixewayflowapi/db/repository/CommentRepository.java
@@ -1,0 +1,7 @@
+package io.mixeway.mixewayflowapi.db.repository;
+
+import io.mixeway.mixewayflowapi.db.entity.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}