+++ backend/src/main/java/io/mixeway/mixewayflowapi/exceptions/FindingNotFoundException.java
@@ -1,0 +1,7 @@
+package io.mixeway.mixewayflowapi.exceptions;
+
+public class FindingNotFoundException extends Exception {
+
+    public FindingNotFoundException() {
+        super("Finding not found");
+    }}