+++ backend/src/main/java/io/mixeway/mixewayflowapi/exceptions/UnauthorizedAccessException.java
@@ -1,0 +1,7 @@
+package io.mixeway.mixewayflowapi.exceptions;
+
+public class UnauthorizedAccessException extends RuntimeException {
+    public UnauthorizedAccessException(String message) {
+        super(message);
+    }
+}