+++ frontend/src/app/model/FindingDTO.ts
@@ -8,6 +8,11 @@ export class FindingDTO {
     inserted: string = "";
     last_seen: string= "";
 }
+interface Comment {
+    inserted: string;
+    message: string;
+    author: string;
+}
 
 export class SingleFindingDTO {
     vulnsResponseDto: FindingDTO | undefined;
@@ -15,5 +20,5 @@ export class SingleFindingDTO {
     recommendation: string = "";
     explanation: string = "";
     refs:string = "";
-
+    comments: Comment[] = [];
 }